Refactoring ColorTile to a component
