+++ lib/Views/Homepage/Components/color_list_tile.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class ColorTile extends StatelessWidget {
+  final int index;
+  const ColorTile({
+    Key? key,
+    required this.index,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Text("$index"),
+    );
+  }
+}
