+++ lib/Views/Homepage/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:testinflutter/Views/Homepage/Components/color_list_tile.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -10,24 +11,9 @@ class HomePage extends StatelessWidget {
         body: Container(
           child: ListView.builder(
               itemCount: 100,
-              itemBuilder: (BuildContext context, index) => ColorListTile(
+              itemBuilder: (BuildContext context, index) => ColorTile(
                     index: index,
                   )),
         ));
-  }
-}
-
-class ColorListTile extends StatelessWidget {
-  final int index;
-  const ColorListTile({
-    Key? key,
-    required this.index,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      title: Text("$index"),
-    );
   }
 }
